Add files via upload
This is the notebook of my house price prediction kaggle competition. I tried to make it in such a way that it is easy to understand. I used plenty of graphs, did plenty of imputations and got a decent dataframe. And my intention is to apply voting regressor with base models RF, XGB, LGBM to get the price.